Added LoggerHelper.cs
