+++ Agero.Core.SplunkLogger/Logger.cs
@@ -1,16 +1,14 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.Linq;
 using System.Net;
-using System.Net.Sockets;
 using System.Threading.Tasks;
 using Agero.Core.Checker;
-using Agero.Core.Lazy;
 using Agero.Core.RestCaller;
 using Agero.Core.RestCaller.Exceptions;
 using Agero.Core.RestCaller.Extensions;
 using Agero.Core.SplunkLogger.Models;
+using Agero.Core.SplunkLogger.Helpers;
 using Newtonsoft.Json;
 
 namespace Agero.Core.SplunkLogger
@@ -19,8 +17,6 @@ namespace Agero.Core.SplunkLogger
     public class Logger : ILogger
     {
         private readonly IReadOnlyDictionary<string, string> _headers;
-
-        private readonly SyncLazy<string> _ec2InstanceId;
 
         /// <summary>Constructor</summary>
         /// <param name="collectorUri">Splunk HTTP collector URL</param>
@@ -47,8 +43,6 @@ namespace Agero.Core.SplunkLogger
                 {
                     {"Authorization", $"Splunk {authorizationToken}"}
                 };
-
-            _ec2InstanceId = new SyncLazy<string>(() => GetAwsMetaDataValue("instance-id"));
         }
 
         /// <summary>Splunk HTTP collector URL</summary>
@@ -66,62 +60,6 @@ namespace Agero.Core.SplunkLogger
         /// <summary>Splunk HTTP collector timeout (milliseconds)</summary>
         public int Timeout { get; }
 
-        private string GetAwsMetaDataValue(string key)
-        {
-            Check.ArgumentIsNullOrWhiteSpace(key, "key");
-
-            try
-            {
-                var uri = new Uri($"http://169.254.169.254/latest/meta-data/{key}");
-                var response = _restCaller.Get(uri, timeout: 100);
-
-                if (response.HttpStatusCode != HttpStatusCode.OK)
-                    return null;
-
-                return response.Text;
-            }
-            catch
-            {
-                return null;
-            }
-        }
-
-        private static string[] GetIpAddresses()
-        {
-            var hostName = Dns.GetHostName();
-            if (string.IsNullOrWhiteSpace(hostName))
-                return new string[] { };
-
-            var hostEntry = Dns.GetHostEntry(hostName);
-
-            return
-                hostEntry.AddressList
-                    .Where(a => a.AddressFamily == AddressFamily.InterNetwork)
-                    .Select(a => a.ToString())
-                    .ToArray();
-        }
-
-        private object GetSystemData()
-        {
-            return
-                new
-                {
-                    userName = Environment.UserName,
-                    userDomainName = Environment.UserDomainName,
-                    operatingSystem = Environment.OSVersion.VersionString,
-                    is64BitOperatingSystem = Environment.Is64BitOperatingSystem,
-                    processorCount = Environment.ProcessorCount,
-                    clrVersion = Environment.Version.ToString(),
-                    is64BitProcess = Environment.Is64BitProcess,
-                    machineName = Environment.MachineName,
-                    localTime = DateTimeOffset.Now,
-                    utcTime = DateTimeOffset.UtcNow,
-                    hostName = Dns.GetHostName(),
-                    ipAddresses = GetIpAddresses(),
-                    ec2InstanceId = _ec2InstanceId.Value
-                };
-        }
-
         private readonly IRESTCaller _restCaller = new RESTCaller();
 
         private string CreateBody(string type, string message, object data, string correlationId)
@@ -129,7 +67,7 @@ namespace Agero.Core.SplunkLogger
             Check.ArgumentIsNullOrWhiteSpace(type, "type");
             Check.ArgumentIsNullOrWhiteSpace(message, "message");
 
-            var body = new LogBody(type, ApplicationName, ApplicationVersion, message, GetSystemData(), data, correlationId);
+            var body = new LogBody(type, ApplicationName, ApplicationVersion, message, LoggerHelper.GetSystemData(), data, correlationId);
 
             return JsonConvert.SerializeObject(body);
         }
@@ -228,7 +166,7 @@ namespace Agero.Core.SplunkLogger
                         url = CollectorUri,
                         applicationName = ApplicationName,
                         applicationVersion = ApplicationVersion,
-                        system = GetSystemData(),
+                        system = LoggerHelper.GetSystemData(),
                         data = data
                     };
                 Trace.WriteLine(JsonConvert.SerializeObject(splunkFailedLog), "ERROR");
